requirements.txt
streamlit>=1.33.0
groq>=0.4.0
deepgram-sdk>=3.0.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
plotly>=5.18.0
streamlit-ace>=0.1.1
pandas>=2.0.0
Pillow>=10.0.0
httpx>=0.25.0
pydub>=0.25.1
gTTS>=2.3.0
